fix(Privilege): add privilege controller fix

+++ src/main/java/org/example/cpas/services/impl/PrivilegeServiceImpl.java
@@ -55,7 +55,9 @@ public class PrivilegeServiceImpl implements PrivilegeService {
 
     @Override
     public void deletePrivilege(Integer privilegeId) {
-        Privilege privilege = this.privilegeRepository.findById(privilegeId).orElseThrow();
-        this.privilegeRepository.delete(privilege);
+        Privilege privilege = this.privilegeRepository.findById(privilegeId).orElse(null);
+        if(privilege!=null){
+            this.privilegeRepository.delete(privilege);
+        }
     }
 }
